requirements.txt
# Core ML and Data Science Libraries
catboost==1.2.8
lightgbm==4.6.0
scikit-learn==1.7.0
numpy==2.2.5
pandas==2.2.3
joblib==1.4.2

# Natural Language Processing
nltk==3.9.1
sentence-transformers==4.1.0

# Hyperparameter Optimization
optuna==4.4.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2

# Web API Framework
fastapi[standard]==0.115.13
uvicorn[standard]>=0.18.0

# Configuration and Logging
pydantic==2.10.3
PyYAML>=6.0
seqlog==0.4.1

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# Security
safety>=2.0.0
bandit>=1.7.0

# Additional utilities
tqdm>=4.64.0
python-dotenv>=1.0.0
